last-week-project/unit-testing-spring-boot/src/main/java/com/herbalife/unittestingspringboot/domain/Car.java
package com.herbalife.unittestingspringboot.domain;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class Car {
    @Value("${model}")
    private String model;
    @Value("${yearOfMake}")
    private int year;

    public String getModel() {
        return model;
    }

    public void setModel(String model) {
        this.model = model;
    }

    public int getYear() {
        return year;
    }

    public void setYear(int year) {
        this.year = year;
    }
}
